Throw an error if the user who is trying to delete a thing doesn't have the appropriate permissions

+++ api/controllers/things/destroy-one-thing.js
@@ -16,11 +16,19 @@ module.exports = {
 
 
   exits: {
-
+    forbidden: {
+      description: 'The user that is making this request doesn\'t have the appropriate permissions to delete this thing.',
+      responseType: 'forbidden'
+    }
   },
 
 
   fn: async function (inputs, exits) {
+    const thing = await Thing.findOne({ id: inputs.id });
+
+    if (thing.owner !== this.req.me.id) {
+      throw 'forbidden';
+    }
 
     await Thing.destroy({ id: inputs.id });
     return exits.success();
